refactor: streamline mutation functions and enhance route organization
