+++ web/src/platform/routes/Auth/VerifyEmail.tsx
@@ -40,8 +40,6 @@ export default function VerifyEmail() {
     const navigate = useNavigate();
     const queryClient = useQueryClient();
     const token = useFragmentToken(REGISTRATION_TOKEN_KEY);
-    const [accountExists, setAccountExists] = useState(false);
-    const [setupMismatch, setSetupMismatch] = useState(false);
     const [lastVerifiedSetup, setLastVerifiedSetup] = useState<RegistrationSetup | null>(null);
     const schema = z.object({
         name: z.string().trim().min(1, 'Name is required').max(127, 'Name cannot exceed 127 characters'),
@@ -100,13 +98,6 @@ export default function VerifyEmail() {
             if (error instanceof ApiError && error.code === 'VERIFY_USER_BAD_TOKEN') {
                 verification.mutate('');
             }
-            if (error instanceof ApiError && error.code === 'REGISTER_SETUP_MISMATCH') {
-                setSetupMismatch(true);
-            }
-            if (error instanceof ApiError && error.code === 'REGISTER_USER_ALREADY_EXISTS') {
-                setAccountExists(true);
-            }
-
             const message =
                 error instanceof ApiError && error.code === 'REGISTER_USER_ALREADY_EXISTS'
                     ? 'An account with this email already exists. Sign in or reset your password to continue.'
@@ -128,6 +119,9 @@ export default function VerifyEmail() {
     const recoverySearch = recoverySetup?.email ? `?${new URLSearchParams({ email: recoverySetup.email })}` : '';
     const recoveryRegisterHref = `/auth/register${recoverySearch}`;
     const recoverySignInHref = `/organizations${recoverySearch}`;
+    const accountExists =
+        completion.error instanceof ApiError && completion.error.code === 'REGISTER_USER_ALREADY_EXISTS';
+    const setupMismatch = completion.error instanceof ApiError && completion.error.code === 'REGISTER_SETUP_MISMATCH';
 
     // Keep transient verification failures retryable while expired credentials remain terminal.
     if (verification.error) {
@@ -146,12 +140,7 @@ export default function VerifyEmail() {
                 <Stack gap={3}>
                     <Banner status="error" title="LongLink could not verify this registration link." />
                     {invalidToken ? null : (
-                        <Button
-                            isLoading={verification.isPending}
-                            label="Retry"
-                            onClick={() => verification.mutate(token)}
-                            variant="primary"
-                        />
+                        <Button label="Retry" onClick={() => verification.mutate(token)} variant="primary" />
                     )}
                     <Button href={recoveryRegisterHref} label="Request a new registration link" />
                 </Stack>
